extract login from app.svelte to login.svelte
